+++ src/App.css
@@ -4,20 +4,6 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  justify-content: space-between;
+  justify-content: flex-start;
   overflow-y: auto;
-}
-
-.App-header {
-  background-color: rgb(133, 86, 107);
-  height: 50px;
-  width: 100%;
-  padding: 20px;
-  color: white;
-  display: flex;
-  justify-content: center;
-}
-
-.App-title {
-  font-size: 1.5em;
 }
