+++ src/components/input-wrapper/style.css
@@ -9,7 +9,7 @@
 }
 
 .input-wrapper-input {
-  border: 1px solid #888;
+  border: 1px solid rgb(223, 217, 217);
   border-radius: 3px;
   width: 55%;
 }
