+++ src/modules/header/style.css
@@ -1,0 +1,12 @@
+.header-container {
+  height: 70px;
+  width: 100%;
+  padding-left: 50px;
+  color: white;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  border: 0.5px solid rgb(223, 217, 217);
+  background-color: rgb(256, 256, 256);
+  box-shadow: 1.5px 3px 5px #d8d8d8;
+}
