+++ src/modules/login/style.css
@@ -1,0 +1,28 @@
+.login-form {
+  margin-top: 50px;
+  width: 600px;
+  height: 300px;
+  box-shadow: 3px 5px 7px #d8d8d8;
+  border: 1px solid rgb(223, 217, 217);
+  border-radius: 3px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  background-color: rgb(256, 356, 256);
+  color: #333;
+}
+
+.login-button-resume {
+  padding: 10px;
+  background-color: rgb(105, 177, 134);
+  width: 300px;
+  color: #eee;
+  border-radius: 3px;
+  font-size: 20px;
+}
+
+.login-button-resume:hover {
+  background-color: rgb(55, 218, 64);
+  color: #fff;
+}
