+++ src/modules/logo/style.css
@@ -1,0 +1,11 @@
+.logo-container {
+  display: flex;
+  height: 50px;
+  color: rgba(105, 155, 84, 0.8);
+  font-size: 30px;
+  font-family: 'Lobster', cursive;
+}
+
+.logo-icon {
+  margin-right: 5px;
+}
