--- src/modules/registration/registration-header/index.js
@@ -1,17 +1,0 @@
-import React, { Component } from 'react';
-import './style.css';
-
-class RegistrationHeader extends Component {
-
-  render() {
-    return (
-
-      <header className="App-header">
-        <h1 className="App-title">Chatilingus</h1>
-      </header>
-    );
-  }
-};
-
-
-export default RegistrationHeader;
