+++ src/modules/registration/style.css
@@ -1,22 +1,15 @@
-.registration-container {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
 .registration-form {
   margin-top: 50px;
   width: 600px;
   height: 400px;
-  box-shadow: 3px 5px 7px #afaeae;
-  border: 1px solid #777;
+  box-shadow: 3px 5px 7px #d8d8d8;
+  border: 1px solid rgb(223, 217, 217);
   border-radius: 3px;
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
+  background-color: rgb(256, 356, 256);
   color: #333;
 }
