+++ src/modules/welcome-page/style.css
@@ -1,0 +1,8 @@
+.welcome-page-container {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+}
